backend_flask/requirements/common.txt
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
PyYAML==5.3.1
python-dotenv==0.13.0
pycryptodome==3.9.7
